Youssef

+++ src/storageManager/FileManager.java
@@ -1,20 +1,20 @@
 package storageManager;
 import java.io.*;
+import javaapplication19.SudokuBoard;
 public class FileManager {
-    public static void saveBoard(String folder, int[][] board) throws IOException {
+    public static void saveBoard(String folder, SudokuBoard board) throws IOException {
         new File(folder).mkdirs();    
         BufferedWriter writer = new BufferedWriter(new FileWriter(folder + "/game.txt"));
         for (int i = 0; i < 9; i++) {
             for (int j = 0; j < 9; j++) {
-                writer.write(board[i][j] + " ");
+                writer.write(board.getGrid()[i][j]+ " ");//getGrid for return int[][]
             }
             writer.newLine();
         }
         writer.close();
     }
-    public static int[][] loadBoard(String folder) throws IOException {
+    public static SudokuBoard loadBoard(String folder) throws IOException {
         int[][] board = new int[9][9];
-        
         BufferedReader reader = new BufferedReader(new FileReader(folder + "/game.txt"));
         for (int i = 0; i < 9; i++) {
             String[] values = reader.readLine().trim().split("\\s+");
@@ -22,14 +22,20 @@ public class FileManager {
                 board[i][j] = Integer.parseInt(values[j]);
             }
         }
-        reader.close();
-        
-        return board;
+        reader.close(); 
+        return new SudokuBoard(board);//constructor for SudokuBoard
     }
     public static boolean exists(String folder) {
         return new File(folder + "/game.txt").exists();
     }
     public static void delete(String folder) {
-        new File(folder + "/game.txt").delete();
+        if(exists(folder))
+        {          
+            new File(folder + "/game.txt").delete();
+        }
+        else
+        {
+            System.out.println("error deleting game");
+        }
     }
 }
